readtags: move the main driver code to readtags-cmd.c
readtags.[ch] has been assumed a kind of library.  So the
compatibility of API has been important.  Therefore aggressive
improvement of readtags command should be done in separated place .
